Fisiologia/python-serial-realtime-app/websocket_server/requirements.txt
# Servidor WebSocket para visualización de señales PPG en tiempo real
# 
# Instalación:
#   pip install -r requirements.txt

# Servidor FastAPI y WebSocket
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Procesamiento de señales
numpy>=1.21.0
scipy>=1.7.0

# Comunicación serial
pyserial>=3.5

# Utilidades
python-multipart>=0.0.6  # Para upload de archivos
pydantic>=2.0.0  # Validación de datos
